requirements_python2.txt
lockfile
python-daemon
appdirs
urllib3==1.24
unittest-xml-reporting
PyYAML
inquirer
future
colorama
pyjwt
click==7.0
requests==2.21.0
tabulate
socketIO_client
mock
enum
